2D Models in SimpleABM Example1

Step 1: Create Agents and Model

Step 2: Initialise the model

Step 3: Run the model

Step 4: Animation and Data

API

Functions for creating agents

Functions for defining model

Functions for initialising, running and visualising

Functions for accessing data.

Functions for creating and modifying a graph

Helper functions for agents

Functions for getting neighbor agents.

Helper functions for patches, nodes, edges

EasyABM.jl

Package Features

API

Functions for creating agents

Functions for defining model

Functions for initialising, running and visualising

Functions for accessing data.

Functions related to graphs

Helper functions for agents

Functions for getting neighbor agents.

Helper functions for patches, nodes, edges

Index

EasyABM.jl

Abmot is an easy to use and performant agent-based modeling tool.

Package Features

API

EasyABM provides following functions for agent based simulations.

Functions for creating agents

Missing docstring.

Missing docstring for create_2d_agent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_2d_agents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_3d_agent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_3d_agents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_graph_agent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_graph_agents. Check Documenter's build log for details.

Functions for defining model

Missing docstring.

Missing docstring for create_2d_model. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_3d_model. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_graph_model. Check Documenter's build log for details.

Functions for initialising, running and visualising

Missing docstring.

Missing docstring for init_model!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for run_model!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for animate_sim. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_interactive_app. Check Documenter's build log for details.

Functions for accessing data.

Missing docstring.

Missing docstring for get_agent_data. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_patch_data. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_node_data. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_edge_data. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_model_data. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_nums_agents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_nums_patches. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_nums_nodes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_nums_edges. Check Documenter's build log for details.

Functions related to graphs

EasyABM.adjacency_matrix — Function

adjacency_matrix(g::EasyABM.SimplePropGraph) -> SparseArrays.SparseMatrixCSC{Int64, Int64}

Returns Adjacency matrix as a Sparse Matrix for a given simple prop graph.

adjacency_matrix(g::EasyABM.DirPropGraph) -> SparseArrays.SparseMatrixCSC{Int64, Int64}

Returns Adjacency matrix as a Sparse Matrix for a given directed prop graph.

Missing docstring.

Missing docstring for create_simple_graph. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_dir_graph. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_node!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for kill_node!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_edge!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for kill_edge!. Check Documenter's build log for details.

Helper functions for agents

Missing docstring.

Missing docstring for get_grid_loc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_node_loc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_id. Check Documenter's build log for details.

Missing docstring.

Missing docstring for agents_at. Check Documenter's build log for details.

Missing docstring.

Missing docstring for num_agents_at. Check Documenter's build log for details.

Missing docstring.

Missing docstring for agent_with_id. Check Documenter's build log for details.

Missing docstring.

Missing docstring for is_alive. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_agents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for num_agents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for kill_agent!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_agent!. Check Documenter's build log for details.

Functions for getting neighbor agents.

Missing docstring.

Missing docstring for neighbors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for in_neighbors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for out_neighbors. Check Documenter's build log for details.

Helper functions for patches, nodes, edges

Missing docstring.

Missing docstring for is_occupied. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_nodeprop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_edgeprop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_nodeprops!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_edgeprops!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_patchprop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_patchprops!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for neighbor_nodes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for neighbor_patches. Check Documenter's build log for details.

Missing docstring.

Missing docstring for in_neighbor_nodes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for out_neighbor_nodes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_nodes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for num_nodes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_edges. Check Documenter's build log for details.

Missing docstring.

Missing docstring for num_edges. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_patches. Check Documenter's build log for details.

Missing docstring.

Missing docstring for num_patches. Check Documenter's build log for details.

Missing docstring.

Missing docstring for random_empty_node. Check Documenter's build log for details.

Missing docstring.

Missing docstring for random_empty_patch. Check Documenter's build log for details.

Index

EasyABM.add_agent!

EasyABM.add_node!

EasyABM.adjacency_matrix

EasyABM.agent_with_id

EasyABM.agents_at

EasyABM.animate_sim

EasyABM.create_2d_agent

EasyABM.create_2d_agents

EasyABM.create_2d_model

EasyABM.create_3d_agent

EasyABM.create_3d_agents

EasyABM.create_3d_model

EasyABM.create_dir_graph

EasyABM.create_edge!

EasyABM.create_graph_agent

EasyABM.create_graph_agents

EasyABM.create_graph_model

EasyABM.create_interactive_app

EasyABM.create_simple_graph

EasyABM.get_agent_data

EasyABM.get_agents

EasyABM.get_edge_data

EasyABM.get_edgeprop

EasyABM.get_edges

EasyABM.get_grid_loc

EasyABM.get_id

EasyABM.get_model_data

EasyABM.get_node_data

EasyABM.get_node_loc

EasyABM.get_nodeprop

EasyABM.get_nodes

EasyABM.get_nums_agents

EasyABM.get_nums_edges

EasyABM.get_nums_nodes

EasyABM.get_nums_patches

EasyABM.get_patch_data

EasyABM.get_patches

EasyABM.get_patchprop

EasyABM.in_neighbor_nodes

EasyABM.in_neighbors

EasyABM.init_model!

EasyABM.is_alive

EasyABM.is_occupied

EasyABM.kill_agent!

EasyABM.kill_edge!

EasyABM.kill_node!

EasyABM.neighbor_nodes

EasyABM.neighbor_patches

EasyABM.neighbors

EasyABM.num_agents

EasyABM.num_agents_at

EasyABM.num_edges

EasyABM.num_nodes

EasyABM.num_patches

EasyABM.out_neighbor_nodes

EasyABM.out_neighbors

EasyABM.random_empty_node

EasyABM.random_empty_patch

EasyABM.run_model!

EasyABM.set_edgeprops!

EasyABM.set_nodeprops!

EasyABM.set_patchprops!