API

EasyABM provides following functions for agent based simulations.

Functions for creating agents

EasyABM.create_2d_agent — Function

create_2d_agent(; kwargs...) -> EasyABM.AgentDict2D{Symbol, Any}

Creates a single 2d agent with properties specified as keyword arguments. Following property names are reserved for some specific agent properties - pos : position - vel : velocity - shape : shape of agent - color : color of agent - size : size of agent - orientation : orientation of agent - keepsrecordof : list of properties that the agent records during time evolution.

EasyABM.create_2d_agents — Function

create_2d_agents(n::Int64; kwargs...) -> Vector{EasyABM.AgentDict2D}

Creates a list of n 2d agents with properties specified as keyword arguments.

create_2d_agents(n::Int64, agent::EasyABM.AgentDict2D) -> Vector{EasyABM.AgentDict2D}

Returns a list of n 2d agents all having same properties as agent.

EasyABM.create_3d_agent — Function

create_3d_agent(; kwargs...) -> EasyABM.AgentDict3D{Symbol, Any}

Creates a single 3d agent with properties specified as keyword arguments. Following property names are reserved for some specific agent properties - pos : position - vel : velocity - shape : shape of agent - color : color of agent - size : size of agent - orientation : orientation of agent - keepsrecordof : list of properties that the agent records during time evolution.

EasyABM.create_3d_agents — Function

create_3d_agents(n::Int64; kwargs...) -> Vector{EasyABM.AgentDict3D}

Creates a list of n 3d agents with properties specified as keyword arguments.

create_3d_agents(n::Int64, agent::EasyABM.AgentDict3D) -> Vector{EasyABM.AgentDict3D}

Returns a list of n 3d agents all having same properties as agent.

EasyABM.create_graph_agent — Function

create_graph_agent(; kwargs...) -> EasyABM.AgentDictGr{Symbol, Any}

Creates a single graph agent with properties specified as keyword arguments. Following property names are reserved for some specific agent properties - node : node where the agent is located on the graph. - shape : shape of agent - color : color of agent - size : size of agent - orientation : orientation of agent - keepsrecordof : list of properties that the agent records during time evolution.

EasyABM.create_graph_agents — Function

create_graph_agents(n::Int64; kwargs...) -> Vector{EasyABM.AgentDictGr}

Creates a list of n graph agents with properties specified as keyword arguments.

create_graph_agents(n::Int64, agent::EasyABM.AgentDictGr) -> Vector{EasyABM.AgentDictGr}

Returns a list of n graph agents all having same properties as agent.

Functions for defining model

EasyABM.create_2d_model — Function

create_2d_model(agents::Vector{EasyABM.AgentDict2D}; graphics, fix_agents_num, grid_size, periodic, random_positions, kwargs...) -> Union{EasyABM.GridModel2D{MortalType}, EasyABM.GridModel2D{StaticType}}

Creates a model with

agents : list of agents.

graphics : if true properties of pos, shape, color, orientation will be assigned to each agent by default, if not already assigned by the user.

fixagentnum : Set it to true if agents do not die and new agents are not born during simulation. If set to false, each agent is

assigned default properties _birth_time, _death_time, _active which are for internal use in the package and must not be modified by the user.

grid_size : A tuple (dimx, dimy) which tells the number of blocks the space is to be divided into along x and y directions. An agent can take

positions from 0 to dimx in x-direction and 0 to dimy in y direction in order to stay within grid space. The word grid in the function create_grid_model does not imply that agents will be restricted to move in discrete steps. The agents can move continuously or in discrete steps depending upon how user implements the step rule. Each grid block is called a patch which like agents can be assigned its own properties. Other than the number of patches in the model, grid_size will also restrict the domain of neighbors function (which when called with either :chessboard or :euclidean metric option) will only take into account the agents within the grid dimensions and will ignore any agents which have crossed the boundary of grid space(unless periodic is set to true).

periodic : If periodic is true the grid space will be periodic in both x and y directions.

random_positions : If this property is true, each agent, which doesn't already have a position defined, will be given a default random continous position.

kwargs : Keyword argments used as model parameters.

EasyABM.create_3d_model — Function

create_3d_model(agents::Vector{EasyABM.AgentDict3D}; graphics, fix_agents_num, grid_size, periodic, random_positions, kwargs...) -> Union{EasyABM.GridModel3D{MortalType}, EasyABM.GridModel3D{StaticType}}

Creates a model with

agents : list of agents.

graphics : if true properties of pos, shape, color, orientation will be assigned to each agent by default, if not already assigned by the user.

fixagentnum : Set it to true if agents do not die and new agents are not born during simulation. If set to false, each agent is

assigned default properties _birth_time, _death_time, _active which are for internal use in the package and must not be modified by the user.

grid_size : A tuple (dimx, dimy, dimz) which tells the number of blocks the space is to be divided into along x, y and z directions. An agent can take

positions from 0 to dimx in x-direction, 0 to dimy in y direction and 0 to dimz in z direction in order to stay within grid space. The word grid in the function create_grid_model does not imply that agents will be restricted to move in discrete steps. The agents can move continuously or in discrete steps depending upon how user implements the step rule. Each grid block is called a patch which like agents can be assigned its own properties. Other than the number of patches in the model, grid_size also restricts the domain of neighbors function (which when called with either :chessboard or :euclidean metric option) will only take into account the agents within the grid dimensions and will ignore any agents which have crossed the boundary of grid space(unless periodic is set to true).

periodic : If periodic is true the grid space will be periodic in x, y and z directions.

random_positions : If this property is true, each agent, which doesn't already have a position defined, will be given a default random continous position.

kwargs : Keyword argments used as model parameters.

EasyABM.create_graph_model — Function

create_graph_model(agents::Vector{EasyABM.AgentDictGr}, graph::Union{Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Graphs.SimpleGraphs.SimpleGraph{Int64}, EasyABM.DirPropGraph, EasyABM.SimplePropGraph}; fix_agents_num, static_graph, graphics, random_positions, kwargs...) -> EasyABM.GraphModel

Creates a model with

agents : list of agents.

graph : A graph created with Graph.jl or with SimpleABM graph functionality.

graphics : if true properties of pos, shape, color, orientation will be assigned to each agent by default, if not already assigned by user.

fixagentnum : Set it to true if agents do not die and new agents are not born during simulation.

static_graph : Set it to false if graph topology needs to be changed during simulation.

random_positions : If this property is true, each agent, which doesn't already have a node property defined, will be given a default random node on the graph.

kwargs : Keyword argments used as model parameters.

Functions for initialising, running and visualising

EasyABM.init_model! — Function

init_model!(model::EasyABM.GridModel2D; initialiser)

Initiates the simulation with a user defined initialiser function which takes the model as its only argument. Model parameters along with agent properties can be set (or modified if set through the create_2d_agents and create_2d_model functions) from within a user defined function and then sending it as initialiser argument in init_model!.

init_model!(model::EasyABM.GridModel3D; initialiser)

Initiates the simulation with a user defined initialiser function which takes the model as its only argument. Model parameters along with agent properties can be set (or modified if set through the create_3d_agents and create_3d_model functions) from within a user defined function and then sending it as initialiser argument in init_model!.

init_model!(model::EasyABM.GraphModel; initialiser)

Initiates the simulation with a user defined initialiser function which takes the model as its only argument. Model parameters along with agent and graph properties can be set (or modified if set through the create_graph_agents and create_graph_model functions) from within a user defined function and then sending it as initialiser argument in init_model!.

EasyABM.run_model! — Function

run_model!(model::EasyABM.GridModel2D; steps, step_rule, aprops, pprops, mprops, save_to_folder) -> Int64

Runs the simulation for steps number of model-steps. Agent properties specified in aprops will replace each agents keeps_record_of list and will be recorded during simulation. The patch properties to be recorded can be specified as list of symbols pprops. The model properties to be recorded are specified as mprops.

run_model!(model::EasyABM.GridModel3D; steps, step_rule, aprops, pprops, mprops, save_to_folder) -> Int64

Runs the simulation for steps number of model-steps. Agent properties specified in aprops will replace each agents keeps_record_of list and will be recorded during simulation. The patch properties to be recorded can be specified as list of symbols pprops. The model properties to be recorded are specified as mprops.

run_model!(model::EasyABM.GraphModel; steps, step_rule, aprops, nprops, eprops, mprops, save_to_folder) -> Int64

Runs the simulation for steps number of model-steps. Agent properties specified in aprops will replace each agents keeps_record_of list and will be recorded during simulation. The patch properties to be recorded can be specified as list of symbols pprops. The model properties to be recorded are specified as mprops.

EasyABM.animate_sim — Function

animate_sim(model::EasyABM.GridModel2D) -> Widgets.Widget{:default, Any}
animate_sim(model::EasyABM.GridModel2D, frames::Int64; plots, path, show_grid, backend) -> Widgets.Widget{:default, Any}

Creates an animation from the data collected during model run.

animate_sim(model::EasyABM.GridModel3D) -> Union{Nothing, MeshCat.DisplayedVisualizer}
animate_sim(model::EasyABM.GridModel3D, frames::Int64; show_grid) -> Union{Nothing, MeshCat.DisplayedVisualizer}

Creates a 3d animation from the data collected during the model run.

animate_sim(model::EasyABM.GraphModel) -> Widgets.Widget{:default, Any}
animate_sim(model::EasyABM.GraphModel, frames::Int64; plots, path, show_graph, backend) -> Widgets.Widget{:default, Any}

Creates an animation from the data collected during model run.

EasyABM.create_interactive_app — Function

create_interactive_app(model::EasyABM.GridModel2D; initialiser, step_rule, agent_controls, model_controls, plots, path, frames, show_grid, backend) -> Widgets.Widget{:default, Any}

Creates an interactive app for the model.

create_interactive_app(model::EasyABM.GridModel3D; initialiser, step_rule, agent_controls, model_controls, plots, frames, show_grid)

Creates an interactive app for the model.

create_interactive_app(model::EasyABM.GraphModel; initialiser, step_rule, agent_controls, model_controls, plots, path, frames, show_graph, backend) -> Widgets.Widget{:default, Any}

Creates an interactive app for the model.

Functions for accessing data.

EasyABM.get_agent_data — Function

get_agent_data(agent::EasyABM.AbstractPropDict, model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, MortalType}, EasyABM.AbstractGridModel{MortalType}}) -> NamedTuple{(:birthtime, :deathtime, :record), _A} where _A<:Tuple{Any, Any, DataFrames.DataFrame}

get_agent_data(agent::EasyABM.AbstractPropDict, model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, StaticType}, EasyABM.AbstractGridModel{StaticType}}) -> NamedTuple{(:record,), Tuple{DataFrames.DataFrame}}

EasyABM.get_patch_data — Function

get_patch_data(patch, model::EasyABM.AbstractGridModel) -> Union{Nothing, NamedTuple{(:record,), Tuple{DataFrames.DataFrame}}}

EasyABM.get_node_data — Function

get_node_data(node::Int64, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Union{Nothing, NamedTuple{(:birthtime, :deathtime, :record), _A} where _A<:Tuple{Any, Any, DataFrames.DataFrame}}

get_node_data(node::Int64, model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Union{Nothing, NamedTuple{(:record,), Tuple{DataFrames.DataFrame}}}

EasyABM.get_edge_data — Function

get_edge_data(i::Int64, j::Int64, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Union{Nothing, NamedTuple{(:birthtime, :deathtime, :record), _A} where _A<:Tuple{Any, Any, DataFrames.DataFrame}}

get_edge_data(edge, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Union{Nothing, NamedTuple{(:birthtime, :deathtime, :record), _A} where _A<:Tuple{Any, Any, DataFrames.DataFrame}}

get_edge_data(i::Int64, j::Int64, model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Union{Nothing, NamedTuple{(:record,), Tuple{DataFrames.DataFrame}}}

get_edge_data(edge, model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Union{Nothing, NamedTuple{(:record,), Tuple{DataFrames.DataFrame}}}

EasyABM.get_model_data — Function

get_model_data(model::Union{EasyABM.AbstractGraphModel, EasyABM.AbstractGridModel}) -> NamedTuple{(:record,), Tuple{DataFrames.DataFrame}}

EasyABM.get_nums_agents — Function

get_nums_agents(model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, MortalType}, EasyABM.AbstractGridModel{MortalType}}, conditions::Function...; labels, plot_result) -> DataFrames.DataFrame

get_nums_agents(model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, StaticType}, EasyABM.AbstractGridModel{StaticType}}, conditions::Function...; labels, plot_result) -> DataFrames.DataFrame

EasyABM.get_nums_patches — Function

get_nums_patches(model::EasyABM.AbstractGridModel, conditions::Function...; labels, plot_result) -> DataFrames.DataFrame

EasyABM.get_nums_nodes — Function

get_nums_nodes(model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}, conditions::Function...; labels, plot_result) -> DataFrames.DataFrame

get_nums_nodes(model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}, conditions::Function...; labels, plot_result) -> DataFrames.DataFrame

EasyABM.get_nums_edges — Function

get_nums_edges(model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}, conditions::Function...; labels, plot_result) -> DataFrames.DataFrame

get_nums_edges(model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}, conditions::Function...; labels, plot_result) -> DataFrames.DataFrame

Functions for creating and modifying a graph

EasyABM.create_simple_graph — Function

create_simple_graph(n::Int64; gtype) -> Union{Nothing, EasyABM.SimplePropGraph{StaticType}}

Creates a simple prop graph with n vertices.

create_simple_graph(structure::Dict{Int64, Vector{Int64}}; gtype) -> Union{Nothing, EasyABM.SimplePropGraph{StaticType}}

Creates a simple prop graph with given structure.

create_simple_graph(A::Matrix{Int64}; gtype) -> Union{Nothing, EasyABM.SimplePropGraph{StaticType}}

Creates a simple prop graph for given Adjacency matrix.

create_simple_graph(A::SparseArrays.SparseMatrixCSC{Int64, Int64}; gtype) -> Union{Nothing, EasyABM.SimplePropGraph{StaticType}}

Creates a simple prop graph for adjacency_matrix given as a Sparse Matrix.

create_simple_graph(g::Graphs.SimpleGraphs.SimpleGraph{Int64}; gtype)

Creates a simple prop graph from a given simple graph created with Graphs.jl.

EasyABM.create_dir_graph — Function

create_dir_graph(n::Int64; gtype) -> EasyABM.DirPropGraph{StaticType}

Creates a directed prop graph with n vertices.

create_dir_graph(in_structure::Dict{Int64, Vector{Int64}}; gtype) -> Union{Nothing, EasyABM.DirPropGraph{StaticType}}

Creates a directed prop graph with given structure.

create_dir_graph(A::Matrix{Int64}; gtype) -> Union{Nothing, EasyABM.DirPropGraph{StaticType}}

Creates a directed prop graph for given Adjacency matrix.

create_dir_graph(A::SparseArrays.SparseMatrixCSC{Int64, Int64}; gtype) -> Union{Nothing, EasyABM.DirPropGraph{StaticType}}

Creates a directed prop graph for adjacency matrix given as a Sparse Matrix.

create_dir_graph(g::Graphs.SimpleGraphs.SimpleDiGraph{Int64}; gtype)

Creates a directed prop graph for a given directed graph created with Graphs.jl.

EasyABM.add_node! — Function

add_node!(node, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}; kwargs...)

Adds a node with properties specified in kwargs to the model's graph.

EasyABM.kill_node! — Function

kill_node!(node, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Any

Removes a node from model graph.

EasyABM.create_edge! — Function

create_edge!(i, j, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}; kwargs...)

Adds an edge with properties kwargs to model graph.

create_edge!(edge, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}; kwargs...)

Adds an edge with properties kwargs to model graph.

EasyABM.kill_edge! — Function

kill_edge!(i, j, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Union{Nothing, Int64}

Removes edge from the model graph.

kill_edge!(edge, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Union{Nothing, Int64}

Removes edge from the model graph.

Helper functions for agents

EasyABM.get_grid_loc — Function

get_grid_loc(agent::EasyABM.AbstractPropDict, model::EasyABM.AbstractGridModel) -> Any

Returns grid location of the agent.

EasyABM.get_node_loc — Function

get_node_loc(agent::EasyABM.AbstractPropDict, model::EasyABM.AbstractGraphModel) -> Any

Returns node location of the agent.

EasyABM.get_id — Function

get_id(agent::EasyABM.AbstractPropDict) -> Any

Returns agents id.

EasyABM.agents_at — Function

agents_at(patch, model::EasyABM.AbstractGridModel) -> Any

Returns list of agents at a given patch.

agents_at(node, model::EasyABM.AbstractGraphModel) -> Any

Returns list of agents at a given node.

EasyABM.num_agents_at — Function

num_agents_at(patch, model::EasyABM.AbstractGridModel) -> Any

Returns number of agents at a given patch.

num_agents_at(node, model::EasyABM.AbstractGraphModel) -> Any

Returns number of agents at a given node.

EasyABM.agent_with_id — Function

agent_with_id(i, model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, MortalType}, EasyABM.AbstractGridModel{MortalType}}) -> Any

Returns agent having given id.

agent_with_id(i, model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, StaticType}, EasyABM.AbstractGridModel{StaticType}}) -> Any

Returns agent having given id.

EasyABM.is_alive — Function

is_alive(agent::EasyABM.AbstractPropDict) -> Any

Returns true if agent is alive else returns false.

EasyABM.get_agents — Function

get_agents(model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, MortalType}, EasyABM.AbstractGridModel{MortalType}}) -> Any
get_agents(model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, MortalType}, EasyABM.AbstractGridModel{MortalType}}, condition::Function) -> Any

get_agents(model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, StaticType}, EasyABM.AbstractGridModel{StaticType}}) -> Any
get_agents(model::Union{EasyABM.AbstractGraphModel{T<:EasyABM.MType, StaticType}, EasyABM.AbstractGridModel{StaticType}}, condition::Function) -> Any

EasyABM.num_agents — Function

num_agents(model::Union{EasyABM.AbstractGraphModel, EasyABM.AbstractGridModel}) -> Any
num_agents(model::Union{EasyABM.AbstractGraphModel, EasyABM.AbstractGridModel}, condition::Function) -> Any

EasyABM.kill_agent! — Function

kill_agent!(agent::EasyABM.AbstractPropDict, model::EasyABM.AbstractGridModel{MortalType}) -> Any

Sets the agent as inactive thus effectively removing from the model. However, the removed agents are permanently removed from the list model.agents only twice in one step i) After the agent_step_function has run for all agents and ii) After the step_rule.

Sets the agent as inactive thus effectively removing from the model. However, the removed agents are permanently removed from the list model.agents only twice in one step i) After the agent_step_function has run for all agents and ii) After the step_rule.

EasyABM.add_agent! — Function

add_agent!(agent, model::EasyABM.GridModel2D{MortalType}) -> Union{Nothing, Int64}

Adds the agent to the model.

add_agent!(agent, model::EasyABM.GridModel3D{MortalType}) -> Union{Nothing, Int64}

Adds the agent to the model.

add_agent!(agent, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{StaticType, MortalType}}) -> Union{Nothing, Int64}

Adds the agent to the model.

Functions for getting neighbor agents.

EasyABM.neighbors — Function

neighbors(agent::EasyABM.AgentDict2D, model::EasyABM.GridModel2D{MortalType}) -> Union{Nothing, Vector{EasyABM.AgentDict2D}}
neighbors(agent::EasyABM.AgentDict2D, model::EasyABM.GridModel2D{MortalType}, dist::Number; metric) -> Union{Nothing, Vector{EasyABM.AgentDict2D}}

Returns active neighboring agents to given agent. If the metric is :chessboard, then with dist =0 only agents present in the current block of the given agent are returned; with dist=1, agents in the current block of the given agent along with agents in the neighbouring 8 blocks are returned; with dist=2 agents in the current block of given agent, along with agents in 24 nearest blocks are returned, and so on. With metric = :euclidean the agents within Euclidean distance dist are returned.

neighbors(agent::EasyABM.AgentDict2D, model::EasyABM.GridModel2D{StaticType}) -> Union{Nothing, Vector{EasyABM.AgentDict2D}}
neighbors(agent::EasyABM.AgentDict2D, model::EasyABM.GridModel2D{StaticType}, dist::Number; metric) -> Union{Nothing, Vector{EasyABM.AgentDict2D}}

Returns active neighboring agents to given agent. If the metric is :chessboard, then with dist =0 only agents present in the current block of the given agent are returned; with dist=1, agents in the current block of the given agent along with agents in the neighbouring 8 blocks are returned; with dist=2 agents in the current block of given agent, along with agents in 24 nearest blocks are returned, and so on. With metric = :euclidean the agents within Euclidean distance dist are returned.

neighbors(agent::EasyABM.AgentDict3D, model::EasyABM.GridModel3D{MortalType}) -> Vector{EasyABM.AgentDict3D}
neighbors(agent::EasyABM.AgentDict3D, model::EasyABM.GridModel3D{MortalType}, dist::Number; metric) -> Vector{EasyABM.AgentDict3D}

Returns active neighboring agents to given agent. If the metric is :chessboard, then with dist =0 only agents present in the current block of the given agent are returned; with dist=1, agents in the current block of the given agent along with agents in the neighbouring 8 blocks are returned; with dist=2 agents in the current block of given agent, along with agents in 24 nearest blocks are returned, and so on. With metric = :euclidean the agents within Euclidean distance dist are returned.

neighbors(agent::EasyABM.AgentDict3D, model::EasyABM.GridModel3D{StaticType}) -> Vector{EasyABM.AgentDict3D}
neighbors(agent::EasyABM.AgentDict3D, model::EasyABM.GridModel3D{StaticType}, dist::Number; metric) -> Vector{EasyABM.AgentDict3D}

Returns active neighboring agents to given agent. If the metric is :chessboard, then with dist =0 only agents present in the current block of the given agent are returned; with dist=1, agents in the current block of the given agent along with agents in the neighbouring 8 blocks are returned; with dist=2 agents in the current block of given agent, along with agents in 24 nearest blocks are returned, and so on. With metric = :euclidean the agents within Euclidean distance dist are returned.

neighbors(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{StaticType, MortalType}}) -> Vector{EasyABM.AgentDictGr}

Returns agents on neighboring nodes of given agent.

neighbors(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{MortalType, StaticType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Vector{EasyABM.AgentDictGr}

Returns agents on neighboring nodes of given agent.

EasyABM.in_neighbors — Function

in_neighbors(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{StaticType, MortalType}}) -> Vector{EasyABM.AgentDictGr}

Returns agents on neighboring incoming nodes of given agent.

in_neighbors(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{MortalType, StaticType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Vector{EasyABM.AgentDictGr}

Returns agents on neighboring incoming nodes of given agent.

EasyABM.out_neighbors — Function

out_neighbors(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{StaticType, MortalType}}) -> Vector{EasyABM.AgentDictGr}

Returns agents on neighboring outgoing nodes of given agent.

out_neighbors(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{MortalType, StaticType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Vector{EasyABM.AgentDictGr}

Returns agents on neighboring outgoing nodes of given agent.

Helper functions for patches, nodes, edges

EasyABM.is_occupied — Function

is_occupied(patch, model::EasyABM.AbstractGridModel) -> Any

Returns true if a patch is occupied.

is_occupied(node, model::EasyABM.AbstractGraphModel) -> Any

Returns true if a node is occupied.

EasyABM.get_nodeprop — Function

get_nodeprop(key::Symbol, node::Int64, model::EasyABM.GraphModel) -> Any

Returns the value for given property name for a node.

EasyABM.get_edgeprop — Function

get_edgeprop(key::Symbol, i::Int64, j::Int64, model::EasyABM.GraphModel) -> Any

Returns the value for given property name for an edge.

get_edgeprop(key::Symbol, edge, model::EasyABM.GraphModel) -> Any

Returns the value for given property name for an edge.

EasyABM.set_nodeprops! — Function

set_nodeprops!(node::Int64, model::EasyABM.GraphModel; kwargs...)

Sets properties of given node.

EasyABM.set_edgeprops! — Function

set_edgeprops!(i::Int64, j::Int64, model::EasyABM.GraphModel; kwargs...)

Sets properties of given edge.

set_edgeprops!(edge, model::EasyABM.GraphModel; kwargs...)

Sets properties of given edge.

EasyABM.get_patchprop — Function

get_patchprop(key, patch, model::EasyABM.AbstractGridModel) -> Any

Returns value of given property of a patch.

EasyABM.set_patchprops! — Function

set_patchprops!(patch, model::EasyABM.AbstractGridModel; kwargs...)

Sets properties of the patch given as keyword arguments.

EasyABM.neighbor_nodes — Function

neighbor_nodes(node::Int64, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Vector{Int64}

Returns nodes neighboring given node.

neighbor_nodes(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Vector{Int64}

Returns nodes neighboring node of the given agent.

neighbor_nodes(node::Int64, model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Vector{Int64}

Returns nodes neighboring given node.

neighbor_nodes(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Vector{Int64}

Returns nodes neighboring node of the given agent.

EasyABM.neighbor_patches — Function

neighbor_patches(agent::EasyABM.AgentDict2D, model::EasyABM.GridModel2D, dist::Int64) -> Vector{Tuple{Int64, Int64}}

Returns patches neighboring the given agent's patch.

neighbor_patches(patch::Tuple{Int64, Int64}, model::EasyABM.GridModel2D, dist::Int64) -> Vector{Tuple{Int64, Int64}}

Returns patches neighboring the given patch.

neighbor_patches(agent::EasyABM.AgentDict3D, model::EasyABM.GridModel3D, dist::Int64) -> Vector{Tuple{Int64, Int64, Int64}}

Returns patches neighboring given agent's patch.

neighbor_patches(patch::Tuple{Int64, Int64}, model::EasyABM.GridModel3D, dist::Int64)

Returning patches neighboring given patch.

EasyABM.in_neighbor_nodes — Function

in_neighbor_nodes(node::Int64, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Vector{Int64}

Returns nodes of incoming edges at given node.

in_neighbor_nodes(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Vector{Int64}

Returns nodes of incoming edges at given agent's node.

in_neighbor_nodes(node::Int64, model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Vector{Int64}

Returns nodes of incoming edges at given node.

in_neighbor_nodes(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Vector{Int64}

Returns nodes of incoming edges at given agent's node.

EasyABM.out_neighbor_nodes — Function

out_neighbor_nodes(node::Int64, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Vector{Int64}

Returns nodes of outgoing edges at given node.

out_neighbor_nodes(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{MortalType, MortalType}, EasyABM.GraphModel{MortalType, StaticType}}) -> Vector{Int64}

Returns nodes of outgoing edges at given agent's node.

out_neighbor_nodes(node::Int64, model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Vector{Int64}

Returns nodes of outgoing edges at given node.

out_neighbor_nodes(agent::EasyABM.AgentDictGr, model::Union{EasyABM.GraphModel{StaticType, MortalType}, EasyABM.GraphModel{StaticType, StaticType}}) -> Vector{Int64}

Returns nodes of outgoing edges at given agent's node.

EasyABM.get_nodes — Function

get_nodes(graph::EasyABM.AbstractPropGraph{MortalType}) -> Any
get_nodes(graph::EasyABM.AbstractPropGraph{MortalType}, condition::Function) -> Any

get_nodes(graph::EasyABM.AbstractPropGraph{StaticType}) -> Any
get_nodes(graph::EasyABM.AbstractPropGraph{StaticType}, condition::Function) -> Any

EasyABM.num_nodes — Function

num_nodes(graph::EasyABM.AbstractPropGraph) -> Any
num_nodes(graph::EasyABM.AbstractPropGraph, condition::Function) -> Any

EasyABM.get_edges — Function

get_edges(graph::EasyABM.AbstractPropGraph{MortalType}) -> Any
get_edges(graph::EasyABM.AbstractPropGraph{MortalType}, condition::Function) -> Any

get_edges(graph::EasyABM.AbstractPropGraph{StaticType}) -> Any
get_edges(graph::EasyABM.AbstractPropGraph{StaticType}, condition::Function) -> Any

EasyABM.num_edges — Function

num_edges(graph::EasyABM.AbstractPropGraph) -> Any
num_edges(graph::EasyABM.AbstractPropGraph, condition::Function) -> Any

EasyABM.get_patches — Function

get_patches(model::EasyABM.GridModel2D) -> Any
get_patches(model::EasyABM.GridModel2D, condition::Function) -> Any

Returns patches satisfying the given condition.

get_patches(model::EasyABM.GridModel3D) -> Any
get_patches(model::EasyABM.GridModel3D, condition::Function) -> Any

Returns patches satisfying given condition.

EasyABM.num_patches — Function

num_patches(model::EasyABM.GridModel2D) -> Any
num_patches(model::EasyABM.GridModel2D, condition::Function) -> Any

Returns number of patches satisfying the given condition.

num_patches(model::EasyABM.GridModel3D) -> Any
num_patches(model::EasyABM.GridModel3D, condition::Function) -> Any

Returns number of patches satisfying given condition.

EasyABM.random_empty_node — Function

random_empty_node(model::EasyABM.AbstractGraphModel) -> Any

Returns an empty node chosen at random. Returns missing if there is no empty node.

EasyABM.random_empty_patch — Function

random_empty_patch(model::EasyABM.GridModel2D) -> Any

Returns a random patch where no agents are present. Rerurns missing if there is no such patch.

random_empty_patch(model::EasyABM.GridModel3D) -> Any

Returns a random patch where no agents are present. Returns missing if there is no such patch.